backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15
psycopg2-binary
python-dotenv
gunicorn
whitenoise
# For development purposes
django-debug-toolbar